add manage listing page

+++ app/Http/Controllers/ListingController.php
@@ -44,6 +44,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store("logos", "public");
         }
 
+        $formFields["user_id"] = auth()->id();
+
         Listing::create($formFields);
 
 
@@ -60,6 +62,11 @@ class ListingController extends Controller
 
     //Update listing data
     public function update(Request $request, Listing $listing) {
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized action");
+        }
 
         $formFields = $request->validate([
             "title" => 'required',
@@ -83,8 +90,19 @@ class ListingController extends Controller
 
         //Delete listing
         public function destroy(Listing $listing) {
+
+            //make sure logged in user is owner
+            if($listing->user_id != auth()->id()) {
+                abort(403, "Unauthorized action");
+            }
+
             $listing->delete();
             return redirect("/")->with("message", "Listing Deleted Successfully");
+        }
+
+        //Manage Listing
+        public function manage() {
+            return view("listings.manage", ["listings" => auth()->user()->listings()->get()]);
         }
 
 }
